Starting on auth stuff. its always more complicated than I expect!

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,24 +14,24 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-//Route::middleware('auth:api')->get('/user', function (Request $request) {
-//    return $request->user();
-//});
+Route::middleware('auth:api')->get('/user', function (Request $request) {
+    return $request->user();
+});
 
-Route::prefix('user')->group(function() {
-    Route::get('/', 'API\UserController@index'); // TODO
-    Route::post('/', 'API\UserController@create'); // TODO
-    Route::get('/{id}', 'API\UserController@get'); // TODO
-    Route::put('/{id}', 'API\UserController@update'); // TODO
-    Route::delete('/{id}', 'API\UserController@delete'); // TODO
+Route::post('login', 'API\AuthController@login');
+Route::post('register', 'API\AuthController@register');
+
+Route::prefix('admin')->group(function() {
+    Route::post('login', 'API/AuthController@adminLogin');
+    Route::post('register', 'API/AuthController@adminRegister');
 });
 
 Route::prefix('sections')->group(function() {
     Route::get('/', 'API\SectionController@index');
-    Route::post('/', 'API\SectionController@create'); // TODO - remember to disallow only-int titles or we gonna hurt later
+    Route::post('/', 'API\SectionController@create')->middleware(['auth:api', 'scope:create-sections']); // TODO - remember to disallow only-int titles or we gonna hurt later
     Route::get('{uniqueIdentifier}', 'API\SectionController@get');
-    Route::put('{id}', 'API\SectionController@update'); // TODO
-    Route::delete('{id}', 'API\SectionController@delete'); // TODO - also consider how to handle cascading
+    Route::put('{id}', 'API\SectionController@update')->middleware(['auth:api', 'scope:edit-sections']); // TODO
+    Route::delete('{id}', 'API\SectionController@delete')->middleware(['auth:api', 'scope:delete-sections']); // TODO - also consider how to handle cascading
     Route::get('{id}/topics', 'API\SectionController@getTopics');
     Route::get('{id}/subsections', 'API\SectionController@subsections');
     Route::get('{id}/parent', 'API\SectionController@parent');
@@ -38,21 +39,21 @@ Route::prefix('sections')->group(function() {
 
 Route::prefix('topics')->group(function() {
     Route::get('/', 'API\TopicController@index');
-    Route::post('/', 'API\TopicController@create');
+    Route::post('/', 'API\TopicController@create')->middleware(['auth:api', 'scope:create']);
     Route::get('{id}', 'API\TopicController@get');
-    Route::put('{id}', 'API\TopicController@update'); // TODO
-    Route::put('{id}/delete', 'API\TopicController@destroy');
+    Route::put('{id}', 'API\TopicController@update')->middleware(['auth:api', 'scope:edit']); // TODO
+    Route::put('{id}/delete', 'API\TopicController@destroy')->middleware(['auth:api', 'scope:delete']);
     Route::get('{id}/posts', 'API\TopicController@getPosts');
     Route::get('{id}/section', 'API\TopicController@section');
 });
 
 Route::prefix('posts')->group(function() {
     Route::get('/', 'API\PostController@index');
-    Route::post('/', 'API\PostController@create');
+    Route::post('/', 'API\PostController@create')->middleware(['auth:api', 'scope:create']);
     Route::get('{id}', 'API\PostController@get');
-    Route::put('{id}', 'API\PostController@update'); // TODO
-    Route::put('{id}/delete', 'API\PostController@destroy');
-    Route::put('{id}', 'API\PostController@unpublish'); // TODO
+    Route::put('{id}', 'API\PostController@update')->middleware(['auth:api', 'scope:edit']); // TODO
+    Route::put('{id}/delete', 'API\PostController@destroy')->middleware(['auth:api', 'scope:delete']);
+    Route::put('{id}', 'API\PostController@unpublish')->middleware(['auth:api', 'scope:delete']); // TODO
     Route::get('{id}/replies', 'API\PostController@getReplies');
 });
